backend/requirements.txt
django==5.0.2
psycopg2==2.9.9
djangorestframework==3.14.0
django-cors-headers==3.14.0
ipython
ipdb
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
